# backend/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": ["message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": ["message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "Subcategory": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "mainCategoryId": {
                        "type": "string"
                    }
                },
                "required": ["id", "name", "mainCategoryId"],
                "type": "object",
                "additionalProperties": false
            },
            "SubcategoryRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "mainCategoryId": {
                        "type": "string"
                    }
                },
                "required": ["name", "mainCategoryId"],
                "type": "object",
                "additionalProperties": false
            },
            "MainCategory": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "expenseTypeId": {
                        "type": "string"
                    },
                    "transactionTypeId": {
                        "type": "string"
                    }
                },
                "required": ["id", "name", "expenseTypeId", "transactionTypeId"],
                "type": "object",
                "additionalProperties": false
            },
            "MainCategoryRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "expenseTypeId": {
                        "type": "string"
                    },
                    "transactionTypeId": {
                        "type": "string"
                    }
                },
                "required": ["name", "expenseTypeId", "transactionTypeId"],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "balanzo-backend",
        "version": "0.0.1",
        "description": "Node.js backend for Balanzo.",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "Roland Csősz"
        }
    },
    "paths": {
...
        "/subcategories": {
            "get": {
                "operationId": "GetSubcategories",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/Subcategory"
                                            },
                                            "type": "array"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Subcategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "CreateSubcategory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Subcategory"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Subcategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubcategoryRequest"
                            }
                        }
                    }
                }
            }
        },
        "/subcategories/{id}": {
            "get": {
                "operationId": "GetSubcategory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Subcategory"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Subcategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateSubcategory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Subcategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubcategoryRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteSubcategory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Subcategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/main_categories": {
            "post": {
                "operationId": "CreateMainCategory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/MainCategory"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["MainCategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MainCategoryRequest"
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "GetMainCategories",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/MainCategory"
                                            },
                                            "type": "array"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["MainCategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            }
        },
        "/main_categories/{id}": {
            "get": {
                "operationId": "GetMainCategory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/MainCategory"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["MainCategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateMainCategory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["MainCategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MainCategoryRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteMainCategory",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["MainCategories"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}